# optimize buffer-memory allocation