--- a/guava/src/com/google/common/io/CharStreams.java
+++ b/guava/src/com/google/common/io/CharStreams.java
@@ -46,15 +46,8 @@ import java.util.List;
 @GwtIncompatible
 public final class CharStreams {
 
-  // 2K chars (4K bytes)
-  private static final int DEFAULT_BUF_SIZE = 0x800;
-
-  /** Creates a new {@code CharBuffer} for buffering reads or writes. */
-  static CharBuffer createBuffer() {
-    return CharBuffer.allocate(DEFAULT_BUF_SIZE);
+  private CharStreams() {
   }
-
-  private CharStreams() {}
 
   /**
    * Copies all characters between the {@link Readable} and {@link Appendable} objects. Does not
@@ -80,14 +73,17 @@ public final class CharStreams {
       checkNotNull(from);
       checkNotNull(to);
       long total = 0;
-      CharBuffer buf = createBuffer();
-      while (from.read(buf) != -1) {
-        buf.flip();
-        to.append(buf);
-        total += buf.remaining();
-        buf.clear();
+
+      try (TransferBuffer<char[]> buffer = TransferBuffer.getCharArrayTransferBuffer()) {
+        CharBuffer buf = CharBuffer.wrap(buffer.get());
+        while (from.read(buf) != -1) {
+          buf.flip();
+          to.append(buf);
+          total += buf.remaining();
+          buf.clear();
+        }
+        return total;
       }
-      return total;
     }
   }
 
@@ -113,14 +109,16 @@ public final class CharStreams {
   static long copyReaderToBuilder(Reader from, StringBuilder to) throws IOException {
     checkNotNull(from);
     checkNotNull(to);
-    char[] buf = new char[DEFAULT_BUF_SIZE];
-    int nRead;
-    long total = 0;
-    while ((nRead = from.read(buf)) != -1) {
-      to.append(buf, 0, nRead);
-      total += nRead;
+    try (TransferBuffer<char[]> buffer = TransferBuffer.getCharArrayTransferBuffer()) {
+      char[] buf = buffer.get();
+      int nRead;
+      long total = 0;
+      while ((nRead = from.read(buf)) != -1) {
+        to.append(buf, 0, nRead);
+        total += nRead;
+      }
+      return total;
     }
-    return total;
   }
 
   /**
@@ -141,14 +139,16 @@ public final class CharStreams {
   static long copyReaderToWriter(Reader from, Writer to) throws IOException {
     checkNotNull(from);
     checkNotNull(to);
-    char[] buf = new char[DEFAULT_BUF_SIZE];
-    int nRead;
-    long total = 0;
-    while ((nRead = from.read(buf)) != -1) {
-      to.write(buf, 0, nRead);
-      total += nRead;
+    try (TransferBuffer<char[]> buffer = TransferBuffer.getCharArrayTransferBuffer()) {
+      char[] buf = buffer.get();
+      int nRead;
+      long total = 0;
+      while ((nRead = from.read(buf)) != -1) {
+        to.write(buf, 0, nRead);
+        total += nRead;
+      }
+      return total;
     }
-    return total;
   }
 
   /**
@@ -239,12 +239,14 @@ public final class CharStreams {
   public static long exhaust(Readable readable) throws IOException {
     long total = 0;
     long read;
-    CharBuffer buf = createBuffer();
-    while ((read = readable.read(buf)) != -1) {
-      total += read;
-      buf.clear();
+    try (TransferBuffer<char[]> buffer = TransferBuffer.getCharArrayTransferBuffer()) {
+      CharBuffer buf = CharBuffer.wrap(buffer.get());
+      while ((read = readable.read(buf)) != -1) {
+        total += read;
+        buf.clear();
+      }
+      return total;
     }
-    return total;
   }
 
   /**